move version helpers to version.go
